Refactor: Move argument building into args
